Move relevant code to log_utils
